% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkTH}
\alias{checkTH}
\title{Determine if Time Homogeneity is present in the give sequence.}
\usage{
checkTH(seq)
}
\arguments{
\item{seq}{- A sequence whose Time Homogeneity is to be determined}
}
\value{
Returns nothing but prints output representing presence or absence of Time Homogeneity
}
\description{
Takes a sequence as input and finds if Time Homogeneity is present or not.
}
\examples{
## Check for a homogenous sequence
seq <- c(1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2)
checkTH(seq)

## Check for a heterogenous sequence
seq <- c(1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,2,1,1,2,2,1,1,2,2,1,1,2,2,1,1,2,2,1,1,2,2,1,1)
checkTH(seq)
}
\references{
[1] Markov Chain Test for Time Dependence and Homogeneity: An Analytical and Empirical Evaluation
Baris Tan and Kamil Yilmaz European Journal of Operational Research 137 (2002) 524-543
}

